FS: Renamed json config identifiers, added web page handler, optimized variable structure

+++ ESP32S2_CocktailCube_V1.3/SystemHelper.h
@@ -38,7 +38,7 @@ class SystemHelper
     String WifiPowerToString(wifi_power_t power);
 
     // Returns the reset reason as string
-    String GetResetReasonString(int cpu);
+    String GetResetReasonString(int8_t cpu);
 
     // Sets the timestamp of the last user action to the current time
     void IRAM_ATTR SetLastUserAction();
